DCW - Version 1.0 - Fully Functioning!
Huge raft of new code to get everything working.

FTP was split out into a subroutine, as we do basically the same
operations for fetching the latest build ID, as we do for fetching the
latest build itself.

This all seems to work on my xubuntu host, and should work with minor
tweaking on OSX.